+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inbox/CTInboxButtonClickListener.java
@@ -26,27 +26,31 @@ class CTInboxButtonClickListener implements View.OnClickListener {
     private final int position;
 
     private ViewPager viewPager;
-    private boolean isBodyClick;
+
+    private final boolean isBodyClick;
+
+    private final int buttonIndex;
 
     CTInboxButtonClickListener(int position, CTInboxMessage inboxMessage, String buttonText, JSONObject jsonObject,
-            CTInboxListViewFragment fragment, boolean isInboxMessageBodyClick) {
+            CTInboxListViewFragment fragment, boolean isInboxMessageBodyClick, int buttonIndex) {
         this.position = position;
         this.inboxMessage = inboxMessage;
         this.buttonText = buttonText;
         this.fragment = fragment; // be sure to pass this as a Weak Ref
         this.buttonObject = jsonObject;
         this.isBodyClick = isInboxMessageBodyClick;
+        this.buttonIndex = buttonIndex;
     }
 
     CTInboxButtonClickListener(int position, CTInboxMessage inboxMessage, String buttonText,
-            CTInboxListViewFragment fragment, ViewPager viewPager, boolean isInboxMessageBodyClick) {
+            CTInboxListViewFragment fragment, ViewPager viewPager, boolean isInboxMessageBodyClick, int buttonIndex) {
         this.position = position;
         this.inboxMessage = inboxMessage;
         this.buttonText = buttonText;
         this.fragment = fragment; // be sure to pass this as a Weak Ref
         this.viewPager = viewPager;
         this.isBodyClick = isInboxMessageBodyClick;
-
+        this.buttonIndex = buttonIndex;
     }
 
 
@@ -66,11 +70,11 @@ class CTInboxButtonClickListener implements View.OnClickListener {
                         }
                     }
 
-                    fragment.handleClick(this.position, buttonText, buttonObject, getKeyValues(inboxMessage),isBodyClick);
+                    fragment.handleClick(this.position, buttonText, buttonObject, getKeyValues(inboxMessage), buttonIndex);
                 }
             } else {
                 if (fragment != null) {
-                    fragment.handleClick(this.position, null, null, null,isBodyClick);
+                    fragment.handleClick(this.position, null, null, null, buttonIndex);
                 }
             }
         }