+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inbox/CTInboxListViewFragment.java
@@ -1,5 +1,7 @@
 package com.clevertap.android.sdk.inbox;
 
+import static com.clevertap.android.sdk.Constants.APP_INBOX_ITEM_INDEX;
+
 import android.content.Context;
 import android.content.Intent;
 import android.graphics.Color;
@@ -41,8 +43,8 @@ public class CTInboxListViewFragment extends Fragment {
 
     interface InboxListener {
 
-        void messageDidClick(Context baseContext, CTInboxMessage inboxMessage, Bundle data,
-                HashMap<String, String> keyValue,boolean isBodyClick);
+        void messageDidClick(Context baseContext, int position, CTInboxMessage inboxMessage, Bundle data,
+                HashMap<String, String> keyValue, int buttonIndex);
 
         void messageDidShow(Context baseContext, CTInboxMessage inboxMessage, Bundle data);
     }
@@ -210,11 +212,11 @@ public class CTInboxListViewFragment extends Fragment {
         }
     }
 
-    void didClick(Bundle data, int position, HashMap<String, String> keyValuePayload, boolean isInboxMessageBodyClick) {
+    void didClick(Bundle data, int position, HashMap<String, String> keyValuePayload, int buttonIndex) {
         CTInboxListViewFragment.InboxListener listener = getListener();
         if (listener != null) {
             //noinspection ConstantConditions
-            listener.messageDidClick(getActivity().getBaseContext(), inboxMessages.get(position), data, keyValuePayload, isInboxMessageBodyClick);
+            listener.messageDidClick(getActivity().getBaseContext(), position, inboxMessages.get(position), data, keyValuePayload, buttonIndex);
         }
     }
 
@@ -265,8 +267,27 @@ public class CTInboxListViewFragment extends Fragment {
         this.mediaRecyclerView = mediaRecyclerView;
     }
 
-    void handleClick(int position, String buttonText, JSONObject jsonObject, HashMap<String, String> keyValuePayload, boolean isInboxMessageBodyClick) {
+    void handleClick(int position, String buttonText, JSONObject jsonObject, HashMap<String, String> keyValuePayload, int buttonIndex) {
+        boolean isInboxMessageButtonClick = jsonObject != null;
+
         try {
+            if (isInboxMessageButtonClick) {
+                String linkType = inboxMessages.get(position).getInboxMessageContents().
+                        get(0).getLinktype(jsonObject);
+                if (linkType.equalsIgnoreCase(Constants.KEY_URL)) {
+                    String actionUrl = inboxMessages.get(position).getInboxMessageContents().get(0)
+                            .getLinkUrl(jsonObject);
+                    if (actionUrl != null) {
+                        fireUrlThroughIntent(actionUrl);
+                    }
+                }
+            } else {
+                String actionUrl = inboxMessages.get(position).getInboxMessageContents().get(0).getActionUrl();
+                if (actionUrl != null) {
+                    fireUrlThroughIntent(actionUrl);
+                }
+            }
+
             Bundle data = new Bundle();
             JSONObject wzrkParams = inboxMessages.get(position).getWzrkParams();
             Iterator<String> iterator = wzrkParams.keys();
@@ -280,26 +301,7 @@ public class CTInboxListViewFragment extends Fragment {
             if (buttonText != null && !buttonText.isEmpty()) {
                 data.putString("wzrk_c2a", buttonText);
             }
-            didClick(data, position, keyValuePayload,isInboxMessageBodyClick);
-            boolean isKVButton = keyValuePayload != null && !keyValuePayload.isEmpty();
-            if (jsonObject != null) {
-                if (isKVButton || inboxMessages.get(position).getInboxMessageContents().get(0).getLinktype(jsonObject)
-                        .equalsIgnoreCase(Constants.COPY_TYPE)) {
-                    //noinspection UnnecessaryReturnStatement
-                    return;
-                } else {
-                    String actionUrl = inboxMessages.get(position).getInboxMessageContents().get(0)
-                            .getLinkUrl(jsonObject);
-                    if (actionUrl != null) {
-                        fireUrlThroughIntent(actionUrl);
-                    }
-                }
-            } else {
-                String actionUrl = inboxMessages.get(position).getInboxMessageContents().get(0).getActionUrl();
-                if (actionUrl != null) {
-                    fireUrlThroughIntent(actionUrl);
-                }
-            }
+            didClick(data, position, keyValuePayload, buttonIndex);
         } catch (Throwable t) {
             Logger.d("Error handling notification button click: " + t.getCause());
         }
@@ -316,7 +318,8 @@ public class CTInboxListViewFragment extends Fragment {
                     data.putString(keyName, wzrkParams.getString(keyName));
                 }
             }
-            didClick(data, position, null,isInboxMessageBodyClick);
+            //pass APP_INBOX_ITEM_INDEX as buttonIndex to indicate the viewPager/item click
+            didClick(data, position, null, APP_INBOX_ITEM_INDEX);
             String actionUrl = inboxMessages.get(position).getInboxMessageContents().get(viewPagerPosition)
                     .getActionUrl();
             fireUrlThroughIntent(actionUrl);